chore: Move all files to tacoq folder instead of src
