--- client_sdks/python/src/tacoq/publisher/__init__.py
@@ -1,3 +1,0 @@
-from .client import PublisherClient
-
-__all__ = ["PublisherClient"]
