--- client_sdks/python/src/tacoq/worker/cli/__init__.py
@@ -1,3 +1,0 @@
-from .cli import cli
-
-__all__ = ["cli"]
