+++ client_sdks/python/tacoq.egg-info/PKG-INFO
@@ -1,0 +1,20 @@
+Metadata-Version: 2.2
+Name: tacoq
+Version: 0.2.2
+Summary: Python SDK to interact with the TacoQ Service API
+Author: Manuel Costa, Pedro Ribeiro
+Classifier: Programming Language :: Python :: 3
+Classifier: License :: OSI Approved :: MIT License
+Classifier: Operating System :: OS Independent
+Requires-Python: >=3.12
+Description-Content-Type: text/markdown
+Requires-Dist: aio-pika>=9.5.3
+Requires-Dist: aiohttp-retry>=2.9.1
+Requires-Dist: aiohttp>=3.11.8
+Requires-Dist: aioresponses>=0.7.8
+Requires-Dist: click>=8.1.7
+Requires-Dist: opentelemetry-api>=1.30.0
+Requires-Dist: pydantic>=2.10.5
+Requires-Dist: pytest-timeout>=2.3.1
+Requires-Dist: uuid>=1.30
+Requires-Dist: watchfiles>=1.0.3
