+++ client_sdks/python/tacoq.egg-info/SOURCES.txt
@@ -1,0 +1,31 @@
+pyproject.toml
+tacoq/__init__.py
+tacoq.egg-info/PKG-INFO
+tacoq.egg-info/SOURCES.txt
+tacoq.egg-info/dependency_links.txt
+tacoq.egg-info/entry_points.txt
+tacoq.egg-info/requires.txt
+tacoq.egg-info/top_level.txt
+tacoq/core/infra/broker/__init__.py
+tacoq/core/infra/broker/client.py
+tacoq/core/infra/broker/config.py
+tacoq/core/infra/relay/__init__.py
+tacoq/core/infra/relay/client.py
+tacoq/core/infra/relay/config.py
+tacoq/core/models/__init__.py
+tacoq/core/models/exception.py
+tacoq/core/models/task.py
+tacoq/core/telemetry/__init__.py
+tacoq/core/telemetry/logger_manager.py
+tacoq/core/telemetry/tracer_manager.py
+tacoq/publisher/__init__.py
+tacoq/publisher/client.py
+tacoq/worker/__init__.py
+tacoq/worker/client.py
+tacoq/worker/config.py
+tacoq/worker/cli/__init__.py
+tacoq/worker/cli/cli.py
+tacoq/worker/cli/importer.py
+tacoq/worker/cli/logger.py
+tacoq/worker/cli/reloader.py
+tacoq/worker/cli/runner.py
