+++ client_sdks/python/tacoq.egg-info/requires.txt
@@ -1,0 +1,10 @@
+aio-pika>=9.5.3
+aiohttp-retry>=2.9.1
+aiohttp>=3.11.8
+aioresponses>=0.7.8
+click>=8.1.7
+opentelemetry-api>=1.30.0
+pydantic>=2.10.5
+pytest-timeout>=2.3.1
+uuid>=1.30
+watchfiles>=1.0.3
