+++ client_sdks/python/tacoq/publisher/__init__.py
@@ -1,0 +1,3 @@
+from tacoq.publisher.client import PublisherClient
+
+__all__ = ["PublisherClient"]
