+++ client_sdks/python/tacoq/worker/cli/__init__.py
@@ -1,0 +1,3 @@
+from tacoq.worker.cli import cli
+
+__all__ = ["cli"]
