+++ client_sdks/python/tests/e2e/test_full.py
@@ -25,13 +25,13 @@ from uuid import uuid4
 
 import pytest
 from opentelemetry.trace import get_current_span
-from src.tacoq.core.infra.broker import BrokerConfig
-from src.tacoq.core.infra.relay import RelayConfig
-from src.tacoq.core.models import Task, TaskInput, TaskOutput, TaskStatus
-from src.tacoq.core.telemetry import LoggerManager, TracerManager
-from src.tacoq.core.telemetry import StructuredMessage as _
-from src.tacoq.publisher import PublisherClient
-from src.tacoq.worker import WorkerApplication, WorkerApplicationConfig
+from tacoq.core.infra.broker import BrokerConfig
+from tacoq.core.infra.relay import RelayConfig
+from tacoq.core.models import Task, TaskInput, TaskOutput, TaskStatus
+from tacoq.core.telemetry import LoggerManager, TracerManager
+from tacoq.core.telemetry import StructuredMessage as _
+from tacoq.publisher import PublisherClient
+from tacoq.worker import WorkerApplication, WorkerApplicationConfig
 
 # =========================================
 # Tasks
