+++ client_sdks/python/tests/unit/manager/test_health.py
@@ -8,7 +8,7 @@ import pytest
 from aiohttp import ClientConnectorError
 from aiohttp.client_reqrep import ConnectionKey
 from aioresponses import aioresponses
-from src.tacoq.core.infra.relay import RelayClient, RelayStates
+from tacoq.core.infra.relay import RelayClient, RelayStates
 
 # =========================================
 # Health Check Tests
