+++ client_sdks/python/tests/unit/manager/test_tasks.py
@@ -10,8 +10,8 @@ from uuid import UUID
 import pytest
 from aiohttp import ClientResponseError
 from aioresponses import aioresponses
-from src.tacoq.core.infra.relay import RelayClient
-from src.tacoq.core.models import Task, TaskStatus
+from tacoq.core.infra.relay import RelayClient
+from tacoq.core.models import Task, TaskStatus
 
 # =========================================
 # Task Retrieval Tests
