+++ client_sdks/python/tests/unit/publisher/test_publisher_client.py
@@ -10,10 +10,10 @@ from unittest import mock
 from uuid import uuid4
 
 import pytest
-from src.tacoq.core.infra.broker import PublisherBrokerClient
-from src.tacoq.core.infra.relay import RelayClient
-from src.tacoq.core.models import Task, TaskStatus
-from src.tacoq.publisher import PublisherClient
+from tacoq.core.infra.broker import PublisherBrokerClient
+from tacoq.core.infra.relay import RelayClient
+from tacoq.core.models import Task, TaskStatus
+from tacoq.publisher import PublisherClient
 
 # =========================================
 # Task Publishing Tests
